requirements.txt
streamlit
qdrant-client
ollama
python-dotenv
pypdf2